Updated TodoItemsController

+++ Models/TodoItemDTO.cs
@@ -2,7 +2,8 @@ namespace TodoApi.Models;
 
 public class TodoItemDTO 
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string? Name { get; set; }
     public bool IsComplete { get; set; }
 }
+
